# src/specs/openapi.yaml
openapi: 3.1.0
info:
  title: Kokoro API
  version: 1.0.0
  description: Kokoro API powers all core features of the web app including user management, friendships, rooms, and chat.
servers:
  - url: https://kokoro-colombe.vercel.app
    description: Production server
  - url: http://localhost:3000
    description: Local development server
paths:
...
  /rooms/{id}:
    patch:
      summary: Update a room
      description: |
        Updates the specified **room**.  
        Only the **host (creator)** of the room can update it.
      operationId: updateRoom
      security:
        - betterAuthSession: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID of the room to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomUpdateInput"
      responses:
        "200":
          description: Room updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        "401":
          description: Unauthorized – The user session is invalid or missing.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: unauthorized
        "404":
          description: Room not found or user is not the host
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Room not found or you are not the host
        "500":
          description: Internal Server Error – Failed to update room.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Failed to update room
    delete:
      summary: Delete a room
      description: |
        Deletes the specified **room** and all its related data.
        Only the **host (creator)** of the room can perform this action.
      operationId: deleteRoom
      security:
        - betterAuthSession: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID of the room to delete
      responses:
        "200":
          description: Room deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        "401":
          description: Unauthorized – The user session is invalid or missing.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: unauthorized
        "404":
          description: Room not found or user is not the host
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Room not found or you are not the host
        "500":
          description: Internal Server Error – Failed to delete room.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Failed to delete room
...
  /chats:
    post:
      summary: Create a new chat
      description: Creates a private or room chat based on the given parameters.
      operationId: createChat
      security:
        - betterAuthSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  enum: [PRIVATE, ROOM]
                memberIds:
                  type: array
                  items:
                    type: string
                name:
                  type: string
              required: [type, memberIds]
      responses:
        "200":
          description: Chat created or existing private chat returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"
        "401":
          description: Unauthorized – The user session is invalid or missing.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: unauthorized
        "500":
          description: Internal Server Error – Failed to create chat.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Failed to create chat
    get:
      summary: Get all chats for current user
      description: Retrieves all chats where the current user is an active member (not deleted).
      operationId: getUserChats
      security:
        - betterAuthSession: []
      responses:
        "200":
          description: List of chats excluding deleted ones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chat"
        "401":
          description: Unauthorized – The user session is invalid or missing.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: unauthorized
        "500":
          description: Internal Server Error – Failed to retrieve chats.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Failed to get chats
  /chats/{chatId}:
    delete:
      summary: Soft delete chat for current user
      description: Marks the chat as deleted for the current user (soft delete from chat members).
      operationId: softDeleteChat
      security:
        - betterAuthSession: []
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Chat soft deleted for the current user
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        "401":
          description: Unauthorized – The user session is invalid or missing.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: unauthorized
        "404":
          description: Chat not found or user is not a member.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Chat not found
        "500":
          description: Internal Server Error – Failed to delete chat.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Failed to delete chat
...
components:
  securitySchemes:
    betterAuthSession:
      type: apiKey
      in: header
      name: Cookie
      description: >
        BetterAuth session cookie (e.g., `better-auth.session=...`).
        This cookie authenticates the user and must be present in all requests requiring authorization.
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        image:
          type: string
          nullable: true
        email:
          type: string
        emailVerified:
          type: boolean
        isOnline:
          type: boolean
          nullable: true
        bio:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        username:
          type: string
          nullable: true
        displayUsername:
          type: string
          nullable: true
...
    Chat:
      type: object
      properties:
        id:
          type: string
          example: "chat_12345"
        type:
          type: string
          enum: [PRIVATE, ROOM]
          example: "ROOM"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        name:
          type: string
          nullable: true
        members:
          type: array
          items:
            $ref: "#/components/schemas/ChatMember"
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
    ChatMember:
      type: object
      properties:
        id:
          type: string
          example: "chat_member_001"
        userId:
          type: string
          example: "user_6789"
        chatId:
          type: string
          example: "chat_12345"
        joinedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          nullable: true
        user:
          $ref: "#/components/schemas/User"
    Message:
      type: object
      properties:
        id:
          type: string
          example: "msg_12345"
        chatId:
          type: string
          example: "chat_12345"
        senderId:
          type: string
          example: "user_6789"
        content:
          type: string
          example: "Hello everyone!"
        imageUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        sender:
          $ref: "#/components/schemas/User"
...
    RoomUpdateInput:
      type: object
      properties:
        name:
          type: string
          example: "New Room Name"
        description:
          type: string
          example: "Updated description"
        isFavorite:
          type: boolean
          example: true
        isActive:
          type: boolean
          example: false
        maxMembers:
          type: integer
          example: 20